Fix duplicate $arch in path.
On the 2 mirrors tested, both listed package paths beginning with the
"ARCH/" (x86/ or x86_64/), so adding $arch to $mirror makes all the
package grabs fail (404).  I changed the code to always set the $arch
variable (empty, "/x86", or "/x86_64"), leave it out of $mirror, and
only use it when grabbing the setup files.